requirements.txt
python-dotenv
selenium == 4.1.0